fix: tests and such
