+++ apps/ensapi/src/omnigraph-api/lib/find-domains/layers/filter-by-name-in.ts
@@ -1,13 +1,13 @@
-import { eq, inArray, sql } from "drizzle-orm";
+import { inArray, sql } from "drizzle-orm";
 import type { InterpretedName } from "enssdk";
 
-import { ensDb, ensIndexerSchema } from "@/lib/ensdb/singleton";
+import { ensDb } from "@/lib/ensdb/singleton";
 
 import { type BaseDomainSet, selectBase } from "./base-domain-set";
 
 /**
- * Filter a base domain set to only canonical Domains whose materialized `canonicalName` exactly
- * matches one of `names`. Validation (max-length, etc.) is enforced at the GraphQL input layer.
+ * Filter a base domain set to Domains whose materialized `canonicalName` exactly matches one of
+ * `names`. Validation (max-length, etc.) is enforced at the GraphQL input layer.
  *
  * Non-canonical rows have `canonicalName = NULL`, so they cannot match by construction — no
  * separate root-anchoring guard is required.
@@ -16,7 +16,8 @@ import { type BaseDomainSet, selectBase } from "./base-domain-set";
  * @param names - Exact InterpretedNames to match against
  */
 export function filterByNameIn(base: BaseDomainSet, names: InterpretedName[]) {
-  // NOTE: empty array in drizzle is a runtime error, so check here
+  // Drizzle footgun: `inArray(col, [])` generates `col in ()`, a Postgres syntax error.
+  // Short-circuit to an explicit empty result.
   if (names.length === 0) {
     return ensDb.select(selectBase(base)).from(base).where(sql`false`).as("baseDomains");
   }
@@ -24,7 +25,6 @@ export function filterByNameIn(base: BaseDomainSet, names: InterpretedName[]) {
   return ensDb
     .select(selectBase(base))
     .from(base)
-    .innerJoin(ensIndexerSchema.domain, eq(ensIndexerSchema.domain.id, base.domainId))
-    .where(inArray(ensIndexerSchema.domain.canonicalName, names))
+    .where(inArray(base.canonicalName, names))
     .as("baseDomains");
 }
